break ClientProvider into pieces
