+++ common/store-api/src/store_api.rs
@@ -13,11 +13,18 @@
 //  limitations under the License.
 //
 
-pub use common_flights::client_provider::StoreClientProvider;
-pub use common_flights::client_provider::TryGetStoreClient;
-pub use common_store_api::store_api::GetStoreApiClient;
-pub use common_store_api::store_api::StoreApis;
-pub use common_store_api::store_api::StoreApisProvider;
-pub use remote_factory::RemoteFactory;
+use std::sync::Arc;
 
-mod remote_factory;
+use crate::MetaApi;
+use crate::StorageApi;
+
+pub trait StoreApis: MetaApi + StorageApi + Send {}
+
+#[async_trait::async_trait]
+pub trait GetStoreApiClient<T>
+where T: StoreApis
+{
+    async fn try_get_store_apis(&self) -> common_exception::Result<T>;
+}
+
+pub type StoreApisProvider<T> = Arc<dyn GetStoreApiClient<T> + Send + Sync>;
