+++ query/src/datasources/util/remote_factory.rs
@@ -17,11 +17,11 @@ use std::sync::Arc;
 
 use common_exception::ErrorCode;
 use common_exception::Result;
+use common_flights::client_provider::StoreClientProvider;
+use common_flights::client_provider::TryGetStoreClient;
 use common_flights::StoreClient;
 
 use crate::configs::Config;
-use crate::datasources::util::StoreClientProvider;
-use crate::datasources::util::TryGetStoreClient;
 
 pub struct RemoteFactory {
     store_client_provider: StoreClientProvider,
